% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_params}
\alias{check_params}
\title{Check parameters for validity. Is automatically run during simulation.}
\usage{
check_params(params, input.df)
}
\arguments{
\item{params}{list of parameters}

\item{input.df}{input data frame}
}
\value{
list of checked parameters and boolean on whether or not to run the simulation
}
\description{
Check parameters for validity. Is automatically run during simulation.
}
\examples{
check_params(params, input.df)
}
