% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_simulation.R
\name{delete_simulation}
\alias{delete_simulation}
\title{Delete simulation only from simulation register}
\usage{
delete_simulation(rootLogDir, simName)
}
\arguments{
\item{rootLogDir}{root logging directory, as set by argument rootLogDir in run_simulation()}

\item{simName}{name of the simulation}
}
\description{
Delete simulation only from simulation register
}
