% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_simulations.R
\name{filter_simulations}
\alias{filter_simulations}
\title{Find simulations based on their parameters}
\usage{
filter_simulations(rootLogDir, ..., condList = NULL)
}
\arguments{
\item{rootLogDir}{logging directory}

\item{...}{}

\item{condList}{list of parameter names and values}
}
\value{
names of simulations that fit the given parameters
}
\description{
Find simulations based on their parameters
}
\examples{
filter_simulations("./logDir", condList = "forgettingRate = 1")
}
