% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Load parameter list saved in a simulation directory}
\usage{
get_params(rootLogDir, simulationName)
}
\arguments{
\item{rootLogDir}{logging directory}

\item{simulationName}{name of the simulation}
}
\description{
Load parameter list saved in a simulation directory
}
