% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Pcols.R
\name{get_Pcols}
\alias{get_Pcols}
\title{Get columns with acoustic features}
\usage{
get_Pcols(input.df)
}
\arguments{
\item{input.df}{input data frame as loaded by load_input_data()}
}
\value{
names of acoustic feature column(s)
}
\description{
Get columns with acoustic features
}
