% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cache.R
\name{load_cache}
\alias{load_cache}
\title{Load cache as data.table}
\usage{
load_cache(logDir, runs, snaps)
}
\arguments{
\item{logDir}{path to simulation}

\item{runs}{run(s) to be loaded, i.e. number or vector of numbers}

\item{snaps}{snapshot(s) to be loaded, i.e. number or vector of numbers}
}
\description{
Load cache as data.table
}
