% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_input_data.R
\name{load_input_data}
\alias{load_input_data}
\title{Load input data as data.table}
\usage{
load_input_data(params)
}
\arguments{
\item{params}{list of params}
}
\description{
Load input data as data.table
}
\examples{
load_input_data(rlist::list.load("./data/params.yaml"))
}
