% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_intLog.R
\name{load_intLog}
\alias{load_intLog}
\title{Load interactions log as data.table}
\usage{
load_intLog(logDir, runs, snaps)
}
\arguments{
\item{logDir}{path to simulation}

\item{runs}{run(s) to be loaded, i.e. number or vector of numbers}

\item{snaps}{snapshot(s) to be loaded, i.e. number or vector of numbers}
}
\description{
Load interactions log as data.table
}
