% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pop.R
\name{load_pop}
\alias{load_pop}
\title{Load population as data table}
\usage{
load_pop(logDir, runs, snaps)
}
\arguments{
\item{logDir}{path to simulation}

\item{runs}{run(s) to be loaded, i.e. number or vector of numbers}

\item{snaps}{snapshot(s) to be loaded, i.e. number or vector of numbers}
}
\description{
Load population as data table
}
