% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_centroids.R
\name{plot_centroids}
\alias{plot_centroids}
\title{Plot the mean of the acoustic features over simulation time}
\usage{
plot_centroids(pop, Pcols, groupVar = c("snapshot", "run"), params)
}
\arguments{
\item{pop}{population data frame as loaded by load_pop()}

\item{Pcols}{string vector of columns in pop with acoustic features as identified by get_Pcols()}

\item{groupVar}{string vector (max. length: 4) of columns in pop which are used to group the data and plot it. Default is c("snapshot", "run"). Further variables are colour-coded or plotted in facets.}

\item{params}{list of parameters as loaded by get_params()}
}
\description{
Plot the mean of the acoustic features over simulation time
}
