% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phonology.R
\name{plot_phonology}
\alias{plot_phonology}
\title{Plot number of sub-phonemes and agreement with canonical phonemes over simulation time}
\usage{
plot_phonology(pop, canonical, params)
}
\arguments{
\item{pop}{population data frame as loaded by load_pop()}

\item{canonical}{name of column in pop (added by the user) which contains the canonical phonemes}

\item{params}{list of parameters as loaded by get_params()}
}
\description{
Plot number of sub-phonemes and agreement with canonical phonemes over simulation time
}
