% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge_simulation.R
\name{purge_simulation}
\alias{purge_simulation}
\title{Purge single simulation from logging directory and simulation register}
\usage{
purge_simulation(rootLogDir, simulationName)
}
\arguments{
\item{rootLogDir}{root logging directory, as set by argument rootLogDir in run_simulation()}

\item{simulationName}{name of the simulation}
}
\description{
Purge single simulation from logging directory and simulation register
}
