% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation_args.R
\name{run_simulation_args}
\alias{run_simulation_args}
\title{Run simulations, using arguments of function as params}
\usage{
run_simulation_args(
  inputDataFile = NULL,
  features = NULL,
  group = NULL,
  label = NULL,
  initial = NULL,
  word = NULL,
  speaker = NULL,
  subsetSpeakers = NULL,
  subsetLabels = NULL,
  createPopulationMethod = "speaker_is_agent",
  bootstrapPopulationSize = 50,
  initialMemoryResampling = FALSE,
  initialMemoryResamplingFactor = 2,
  removeOriginalExemplarsAfterResampling = FALSE,
  productionBasis = "word",
  productionResampling = "SMOTE",
  productionResamplingFallback = "label",
  productionMinTokens = 10,
  productionSMOTENN = 5,
  perceptionModels = "singleGaussian",
  memoryIntakeStrategy = c("mahalanobisDistance", "maxPosteriorProb"),
  mahalanobisProbThreshold = 0.95,
  posteriorProbThr = 1/3,
  perceptionOOVNN = 5,
  computeGMMsInterval = 100,
  purityRepetitions = 5,
  purityThreshold = 0.75,
  forgettingRate = 1,
  interactionPartners = "betweenGroups",
  speakerProb = NULL,
  listenerProb = NULL,
  runMode = "single",
  multipleABMRuns = 3,
  nrOfSnapshots = 1,
  interactionsPerSnapshot = 10,
  rootLogDir = "./logDir",
  notes = ""
)
}
\arguments{
\item{inputDataFile}{}

\item{features}{}

\item{group}{}

\item{label}{}

\item{initial}{}

\item{word}{}

\item{speaker}{}

\item{subsetSpeakers}{}

\item{subsetLabels}{}

\item{createPopulationMethod}{}

\item{bootstrapPopulationSize}{}

\item{initialMemoryResampling}{}

\item{initialMemoryResamplingFactor}{}

\item{removeOriginalExemplarsAfterResampling}{}

\item{productionBasis}{}

\item{productionResampling}{}

\item{productionResamplingFallback}{}

\item{productionMinTokens}{}

\item{productionSMOTENN}{}

\item{perceptionModels}{}

\item{memoryIntakeStrategy}{}

\item{mahalanobisProbThreshold}{}

\item{posteriorProbThr}{}

\item{perceptionOOVNN}{}

\item{computeGMMsInterval}{}

\item{purityRepetitions}{}

\item{purityThreshold}{}

\item{forgettingRate}{}

\item{interactionPartners}{}

\item{speakerProb}{}

\item{listenerProb}{}

\item{runMode}{}

\item{multipleABMRuns}{}

\item{nrOfSnapshots}{}

\item{interactionsPerSnapshot}{}

\item{rootLogDir}{}

\item{notes}{}
}
\description{
Run simulations, using arguments of function as params
}
