% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{u_fronting}
\alias{u_fronting}
\title{Vowels of 22 Standard Southern British English speakers}
\format{
A data frame with 2390 rows and 8 variables:
\describe{
\item{sl_rowIdx}{row index for identification}
\item{speaker}{speaker code}
\item{age}{age group of speaker}
\item{word}{word class}
\item{label}{canonical phoneme associated with the vowel in given word}
\item{DCT0}{first DCT coefficient computed from second formant trajectory}
\item{DCT1}{second DCT coefficient computed from second formant trajectory}
\item{DCT2}{third DCT coefficient computed from second formant trajectory}
}
}
\source{
Jonathan Harrington (email: jmh@phonetik.uni-muenchen.de)
}
\usage{
u_fronting
}
\description{
A dataset containing acoustic data (DCT coefficients) for three vowels /i, u, ju/
produced by 11 older and 11 younger British English speakers.
}
\keyword{datasets}
