% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_params.R
\name{validate_params}
\alias{validate_params}
\title{Check parameters for validity. Is automatically run during simulation.}
\usage{
validate_params(params, input.df)
}
\arguments{
\item{params}{list of parameters}

\item{input.df}{input data frame}
}
\value{
list of checked parameters and boolean on whether or not to run the simulation
}
\description{
Check parameters for validity. Is automatically run during simulation.
}
