% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_simulation.R
\name{delete_simulation}
\alias{delete_simulation}
\title{Delete simulation only from simulation register}
\usage{
delete_simulation(simulationName_, rootLogDir)
}
\arguments{
\item{simulationName_}{name of the simulation}

\item{rootLogDir}{logging directory}
}
\description{
Delete simulation only from simulation register
}
