% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_simulations.R
\name{filter_simulations}
\alias{filter_simulations}
\title{Find simulations based on their parameters}
\usage{
filter_simulations(rootLogDir, ..., condList = NULL)
}
\arguments{
\item{rootLogDir}{logging directory}

\item{...}{any number of arguments, where the argument name must be from the list of model parameters, connected to a value with a logical operator, e.g. parameter == value}

\item{condList}{list of parameter names and values, e.g. parameter = value; default is NULL.}
}
\value{
names of simulations that fit the given parameters
}
\description{
Find simulations based on their parameters
}
