% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_field_from_sim_register.R
\name{get_field_from_sim_register}
\alias{get_field_from_sim_register}
\title{Title}
\usage{
get_field_from_sim_register(rootLogDir, ...)
}
\arguments{
\item{rootLogDir}{root logging directory, as set by argument rootLogDir in run_simulation()}

\item{...}{field names}
}
\value{

}
\description{
Title
}
