% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_N_Pcols.R
\name{get_N_Pcols}
\alias{get_N_Pcols}
\title{Compute number of columns with acoustic features}
\usage{
get_N_Pcols(data)
}
\arguments{
\item{data}{a data frame; usually either input data as loaded by load_input_data() or population as loaded by load_pop()}
}
\value{
number of columns
}
\description{
Compute number of columns with acoustic features
}
