% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Load parameter list saved in a simulation directory}
\usage{
get_params(rootLogDir, simName)
}
\arguments{
\item{rootLogDir}{root logging directory, as set by argument rootLogDir in run_simulation()}

\item{simName}{name of the simulation}
}
\description{
Load parameter list saved in a simulation directory
}
