% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Pcols.R
\name{get_Pcols}
\alias{get_Pcols}
\title{Get columns with acoustic features}
\usage{
get_Pcols(data)
}
\arguments{
\item{data}{a data frame; usually either input data as loaded by load_input_data() or population as loaded by load_pop()}
}
\value{
names of acoustic feature column(s)
}
\description{
Get columns with acoustic features
}
