% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cache.R
\name{load_cache}
\alias{load_cache}
\title{Load cache as data.table}
\usage{
load_cache(logDir, runs, snaps)
}
\arguments{
\item{logDir}{logging directory}

\item{runs}{number of independent parallel runs, i.e. params$multipleABMRuns or 1 in case of single run}

\item{snaps}{number of snapshots, i.e. params$nrOfSnapshots}
}
\description{
Load cache as data.table
}
\examples{
load_cache("ABM20211022152431", 1, 100)
}
