% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_intLog.R
\name{load_intLog}
\alias{load_intLog}
\title{Load interactions log as data.table}
\usage{
load_intLog(logDir, runs, snaps)
}
\arguments{
\item{logDir}{logging directory}

\item{runs}{number of independent parallel runs, i.e. params$multipleABMRuns or 1 in case of single run}

\item{snaps}{number of snapshots, i.e. params$nrOfSnapshots}
}
\description{
Load interactions log as data.table
}
\examples{
load_intLog("ABM20211022152431", 1, 100)
}
