% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pop.R
\name{load_pop}
\alias{load_pop}
\title{Load population as data table}
\usage{
load_pop(logDir, runs, snaps)
}
\arguments{
\item{logDir}{logging directory}

\item{runs}{number of independent parallel runs, i.e. params$multipleABMRuns or 1 in case of single run}

\item{snaps}{number of snapshots, i.e. params$nrOfSnapshots}
}
\description{
Load population as data table
}
