% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge_simulation.R
\name{purge_simulation}
\alias{purge_simulation}
\title{Purge single simulation from logging directory and simulation register}
\usage{
purge_simulation(simulationName, rootLogDir)
}
\arguments{
\item{simulationName}{name of the simulation}

\item{rootLogDir}{logging directory}
}
\description{
Purge single simulation from logging directory and simulation register
}
\examples{
purge_simulation("ABM20211022152431", "./logDir")
}
