% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge_uncompleted_simulations.R
\name{purge_uncompleted_simulations}
\alias{purge_uncompleted_simulations}
\title{Purge uncompleted simulations from logging directory and simulations register}
\usage{
purge_uncompleted_simulations(rootLogDir)
}
\arguments{
\item{rootLogDir}{root logging directory, as set by argument rootLogDir in run_simulation()}
}
\description{
Purge uncompleted simulations from logging directory and simulations register
}
