% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Start the simulation}
\usage{
run_simulation(paramsFile = "inst/extdata/params.yaml")
}
\arguments{
\item{paramsFile}{path to the params.yaml file, default is "data/params.yaml"}
}
\description{
Start the simulation
}
